refactor(services): remove useless module
